+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * First release.
 * List users assigned to a given role. Refs UIROLES-31.
+* Retrieve up to 5000 capabilities at once. Refs UIROLES-47.